Refactor, still needs styling and functionality work
