+++ frontend/pages/admin/UserManagementPage/helpers/userManagementHelpers.ts
@@ -1,12 +1,4 @@
-import {
-  isEmpty,
-  isEqual,
-  isPlainObject,
-  isString,
-  reduce,
-  trim,
-  union,
-} from "lodash";
+import { isEqual } from "lodash";
 
 import { IInvite } from "interfaces/invite";
 import { IUser, IUserUpdateBody, IUpdateUserFormData } from "interfaces/user";
@@ -17,12 +9,6 @@ type ICurrentUserData = Pick<
   IUser,
   "global_role" | "teams" | "name" | "email" | "sso_enabled"
 >;
-
-interface ILocationParams {
-  pathPrefix?: string;
-  routeTemplate?: string;
-  routeParams?: { [key: string]: number };
-}
 
 interface IRoleOptionsParams {
   isPremiumTier?: boolean;
@@ -69,35 +55,6 @@ const generateUpdateData = (
   );
 };
 
-export const getNextLocationPath = ({
-  pathPrefix = "",
-  routeTemplate = "",
-  routeParams = {},
-}: ILocationParams): string => {
-  const pathPrefixFinal = isString(pathPrefix) ? pathPrefix : "";
-  const routeTemplateFinal = (isString(routeTemplate) && routeTemplate) || "";
-  const routeParamsFinal = isPlainObject(routeParams) ? routeParams : {};
-
-  let routeString = "";
-
-  if (!isEmpty(routeParamsFinal)) {
-    routeString = reduce(
-      routeParamsFinal,
-      (string, value, key) => {
-        return string.replace(`:${key}`, encodeURIComponent(value));
-      },
-      routeTemplateFinal
-    );
-  }
-
-  const nextLocation = union(
-    trim(pathPrefixFinal, "/").split("/"),
-    routeString.split("/")
-  ).join("/");
-
-  return `/${nextLocation}`;
-};
-
 export const roleOptions = ({
   isPremiumTier,
   isApiOnly,
@@ -141,6 +98,5 @@ export const roleOptions = ({
 
 export default {
   generateUpdateData,
-  getNextLocationPath,
   roleOptions,
 };
